test(executor): split history normalization and prompt budget tests into init file
